requirements.txt
colorama
python-nmap
requests
urllib3
paramiko
ftplib